% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{available_years}
\alias{available_years}
\title{Get list of years available for covariate calculation}
\usage{
available_years(eden_path = file.path("~/water"), new = FALSE)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}

\item{new}{logical, should \code{available_years} only list years which have changed
since last download}
}
\value{
vector of years
}
\description{
Get list of years available for covariate calculation
}
