% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{default_colony_buffers}
\alias{default_colony_buffers}
\title{Return default colony buffers in m for each species}
\usage{
default_colony_buffers()
}
\value{
named list of colony buffers in m for each species
}
\description{
Return default colony buffers in m for each species
}
