% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edenR-package.R
\docType{package}
\name{edenR}
\alias{edenR-package}
\alias{edenR}
\title{Retrieve and use EDEN data}
\description{
The \strong{edenR} package provides functions to retrieve, process and
summarize the EDEN water depth data. The data begin in 1991 and are continuously updated today.
}
\author{
\strong{Maintainer}: Ethan P. White \email{ethan@weecology.org} (\href{https://orcid.org/0000-0001-6728-7745}{ORCID})

Authors:
\itemize{
  \item Ethan P. White \email{ethan@weecology.org} (\href{https://orcid.org/0000-0001-6728-7745}{ORCID})
  \item Glenda M. Yenni \email{glenda@weecology.org} (\href{https://orcid.org/0000-0001-6969-1848}{ORCID})
}

}
\keyword{internal}
