% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Calculate dry days from everwader}
\usage{
get_boundaries(
  path =
    "https://raw.githubusercontent.com/weecology/EvergladesWadingBird/refs/heads/main/SiteandMethods/",
  level = "subregions",
  colony_buffers = default_colony_buffers()
)
}
\arguments{
\item{path}{path to regions shapefile}

\item{level}{region level to load (all, wcas, or subregions)}
}
\description{
Calculate dry days from everwader
}
