% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{get_eden_covariates}
\alias{get_eden_covariates}
\title{Generate annual scale water covariates using EDEN data}
\usage{
get_eden_covariates(
  level = "subregions",
  eden_path = file.path("~/water"),
  years = available_years(eden_path),
  boundaries_path =
    "https://raw.githubusercontent.com/weecology/EvergladesWadingBird/refs/heads/main/SiteandMethods/regions/"
)
}
\arguments{
\item{level}{region level to load (all, wcas, or subregions)}

\item{eden_path}{path where the EDEN data should be stored}

\item{years}{numeric vector of years to generate covariates for,
defaults to all available years}

\item{boundaries_path}{name of a shape file holding the boundaries
within which to calculate covariates}
}
\value{
data.frame covariate data including columns for region, year,
covariate, value, and the geometry of the region
}
\description{
Generate annual scale water covariates using EDEN data
}
