% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_data.R
\name{get_eden_data}
\alias{get_eden_data}
\title{Downloads new EDEN depth data, calculates covariates, appends to covariate file}
\usage{
get_eden_data()
}
\description{
Downloads new EDEN depth data, calculates covariates, appends to covariate file
}
