% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{get_eden_depths}
\alias{get_eden_depths}
\title{Generate regional daily mean depth using EDEN data}
\usage{
get_eden_depths(
  level = "subregions",
  eden_path = file.path("Water"),
  years = available_years(eden_path),
  boundaries_path =
    "https://raw.githubusercontent.com/weecology/EvergladesWadingBird/refs/heads/main/SiteandMethods/",
  colony_buffers = default_colony_buffers()
)
}
\arguments{
\item{level}{region level to load (all, wcas, or subregions)}

\item{eden_path}{path where the EDEN data should be stored}

\item{years}{numeric vector of years to collect, defaults to all available years}

\item{boundaries_path}{name of a shape file holding the boundaries
within which to calculate depth}
}
\value{
data.frame covariate data including columns for region, date,
depth value, and the geometry of the region
}
\description{
Generate regional daily mean depth using EDEN data
}
