% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_get.R
\name{getstage}
\alias{getstage}
\title{Retrieve stage data from the EDEN Database}
\usage{
getstage(sites = "NMP", start = "2013-06-15", end = "2015-06-15")
}
\arguments{
\item{sites}{character}

\item{start}{date in POSIX}

\item{end}{date in POSIX}
}
\description{
getstage
}
\examples{
\dontrun{#'

\\url{http://sofia.usgs.gov/eden/eve/index.php?timeseries_start=2015-05-07&timeseries_end=2015-06-16&site_list[]=NMP&water_level=stage&day_hour=daily&max=5&dry=dry&hydrograph_query=Update+Selection}

dt<-getstage(sites="NMP",start="2015-05-07",end="2015-06-15")
plot(dt[,1],dt[,2])
dt<-getstage(sites="NMP",start="2013-06-15",end="2015-06-15")
plot(dt[,1],dt[,2])
}
}

