% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{load_boundaries}
\alias{load_boundaries}
\title{Load boundaries (deprecated)}
\usage{
load_boundaries(
  path =
    "https://raw.githubusercontent.com/weecology/EvergladesWadingBird/refs/heads/main/SiteandMethods/",
  level = "subregions",
  colony_buffers = default_colony_buffers()
)
}
\arguments{
\item{path}{path to regions shapefile}

\item{level}{region level to load (all, wcas, or subregions)}
}
\description{
Deprecated. Use \code{get_boundaries()} instead.
}
