% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eden.R
\name{update_last_download}
\alias{update_last_download}
\title{Write new metadata file for files already downloaded}
\usage{
update_last_download(eden_path, metadata)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}

\item{metadata}{EDEN file metadata}
}
\description{
Write new metadata file for files already downloaded
}
