% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_data.R
\name{update_water}
\alias{update_water}
\title{Writes new water data}
\usage{
update_water(eden_path)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}
}
\description{
Writes new water data
}
